move Report ob to diff folder, add map creation in utils, add map creation in main with options for testing and skipping
